requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
loguru==0.7.2
streamlit==1.29.0
watchdog==3.0.0
